requirement.txt
asgiref==3.7.2
Django==5.0.3
django-haystack==3.2.1
sqlparse==0.4.4
typing_extensions==4.10.0
Whoosh==2.7.4
